Issue 218: eliminate some SysDep template params and some warnings

+++ test/internals/multisend/multicast.h
@@ -92,22 +92,22 @@ public:
 		if (task_id == root) {
 			for (x = 0; x < count; ++x) {
 				if (((unsigned *)_source)[x] != 1) {
-					fprintf(stderr, "Corrupted source buffer at index %d. stop.\n", x);
+					fprintf(stderr, "Corrupted source buffer at index %zd. stop.\n", x);
 					break;
 				}
-				if (((unsigned *)_result)[x] != -1) {
-					fprintf(stderr, "Corrupted result buffer at index %d. stop.\n", x);
+				if (((unsigned *)_result)[x] != (unsigned)-1) {
+					fprintf(stderr, "Corrupted result buffer at index %zd. stop.\n", x);
 					break;
 				}
 			}
 		} else {
 			for (x = 0; x < count; ++x) {
 				if (((unsigned *)_source)[x] != 1) {
-					fprintf(stderr, "Corrupted source buffer at index %d. stop.\n", x);
+					fprintf(stderr, "Corrupted source buffer at index %zd. stop.\n", x);
 					break;
 				}
 				if (((unsigned *)_result)[x] != 1) {
-					fprintf(stderr, "Incorrect result at index %d. stop.\n", x);
+					fprintf(stderr, "Incorrect result at index %zd. stop.\n", x);
 					break;
 				}
 			}
